+++ RELEASE_NOTES
@@ -10,6 +10,7 @@
   - StringTemplate no longer shares template groups across threads, to work around concurrency issues
   - added SqliteDatabaseRule for testing convenience
   - DefineStatementLexer handles predicates that look like definitions better
+  - add a TemplateEngine based on Apache Commons Text StringSubstitutor
 
 3.1.1
   - Improve IBM JDK compatibility with default methods
